feat: add CLI token generation route and inspect routes for server and tool management
- Implemented `createCliTokenRoute` to generate CLI tokens for localhost access.
- Added `inspectHelpers` for handling server and tool inspection logic.
- Created `inspectRoutes` to manage server and tool inspection requests.
- Developed `toolRoutes` for listing and invoking tools via the server manager.
